va: move DNS-01 code from va.go to dns.go
